requirements.txt
streamlit
openai
requests
python-dotenv
faiss-cpu
tqdm
numpy
pillow
